Implement Timer and Scramble logic
Created a useTimer hook which exposes an elapsedTime state,
and functions to start and stop the timer. Then added event
listeners and a lock to handle space key event logic.

Created utility functions in utils for creating a new 3x3
scramble, and for formatting the time from ms to a readable string.